Fix NERDTree file highlight on hover in Neovim (#38)
A couple of things happen when a file in NERDTree is hovered:
  1 - the `ColumnLine` highlight is applied to it
  2 - the `NERDTreeFile` highlight is applied to it

NERDTree by default maps the `NERDTreeFile` highlight to the `Normal`
highlight set in a theme. Since `Normal` has been set to have a
particular foreground and background you would get an issue where the
`ColumnLine` highlight is applied (black background for the line) when
you hover a file in NERDTree, and then the `Normal` highlight is applied
which would overwrite the background of the file name text resulting in
a confusing highlight.

This commit adds dedicated colors to `NERDTreeFile` highlight, mainly
just sets the foreground so that the background can be set correctly
depending on the context.

Issue happens in Neovim, not in Vim.